refactor(map): swap to use a proper viewport in the map
WARNING: @Dom - to view you char sheet bits and work on them you'll need to comment the body{ overflow:hidden } in the App css module.
This is required so the interactions (zoom, drag, pinch etc) all work on the viewport currently.

We'll need to work out the best way to display the char sheet.