feat(*): add page to prompt for project details
The project details are stored in local storage, and it doesn't
initialize the firebase config until there's something there. For now,
if the credentials specified are invalid it will just break until you
manually delete the firebaseConfig object in the local storage. This
isn't entirely ideal, but it'll do for now.