feat(assets): create asset now works (see Jeff)
This work got the Create asset panel working again with PC and NPC assets.
When you create one it will appear and be drag+droppable below it.
The Map still needs a-lot of work so I don't end up repeating code over and over, a LayerContainer that can contain shared code is probably the best option.